DAT-12910 changed config to diff/diffChangelog
