+++ src/main/groovy/liquibase/harness/diff/DiffChangelogTests.groovy
@@ -10,10 +10,12 @@ import spock.lang.Specification
 import spock.lang.Unroll
 
 import static liquibase.harness.diff.DiffCommandTestHelper.*
-import static liquibase.harness.util.TestUtils.*
-import static liquibase.harness.util.FileUtils.*
+import static liquibase.harness.util.FileUtils.deleteFile
+import static liquibase.harness.util.FileUtils.getResourceContent
+import static liquibase.harness.util.TestUtils.chooseRollbackStrategy
+import static liquibase.harness.util.TestUtils.executeCommandScope
 
-class DiffCommandTest extends Specification {
+class DiffChangelogTests extends Specification {
     @Shared
     RollbackStrategy strategy
     @Shared
@@ -94,7 +96,6 @@ class DiffCommandTest extends Specification {
                     "Directory for stored object '" + expectedObjectType + "' was not created at path: " + objectDir.getPath() + "!"
         }
 
-
         then: "validate generated diff changelog"
         assert validateGeneratedDiffChangelog(testInput)
 
@@ -110,43 +111,5 @@ class DiffCommandTest extends Specification {
 
         where:
         testInput << buildTestInput()
-    }
-
-    @Unroll
-    def "apply diff test against #testInput.referenceDatabase.name #testInput.referenceDatabase.version to targetDatabase #testInput.targetDatabase.name #testInput.targetDatabase.version"() {
-        given: "read input data for diff test"
-        Map<String, Object> argsMap = new HashMap()
-        argsMap.put("url", testInput.targetDatabase.url)
-        argsMap.put("username", testInput.targetDatabase.username)
-        argsMap.put("password", testInput.targetDatabase.password)
-        argsMap.put("referenceUrl", testInput.referenceDatabase.url)
-        argsMap.put("referenceUsername", testInput.referenceDatabase.username)
-        argsMap.put("referencePassword", testInput.referenceDatabase.password)
-        argsMap.put("changelogFile", testInput.pathToChangelogFile)
-
-        and: "check databases are online"
-        assert testInput.targetDatabase.database.getConnection() instanceof JdbcConnection: "Target database " +
-                "${testInput.targetDatabase.name}${testInput.targetDatabase.version} is offline!"
-        assert testInput.referenceDatabase.database.getConnection() instanceof JdbcConnection: "Reference database " +
-                "${testInput.referenceDatabase.name}${testInput.referenceDatabase.version} is offline!"
-
-        when: "execute update and diff commands"
-
-        executeCommandScope("update", argsMap)
-        argsMap.put("excludeObjects", "(?i)posts, (?i)authors, (?i)databasechangelog, (?i)databasechangeloglock")
-        String generatedDiffContent = removeDatabaseInfoFromDiff(executeCommandScope("diff", argsMap).toString())
-
-        then: "validate generated diff"
-        String expectedDiffContent = removeDatabaseInfoFromDiff(getResourceContent(testInput.pathToExpectedDiffFile))
-        assert expectedDiffContent == generatedDiffContent
-
-        cleanup: "rollback changes"
-        strategy.performRollback(argsMap)
-
-        where:
-        testInput << buildTestInput()
-    }
-    def cleanupSpec() {
-        strategy.cleanupDatabase(databases)
     }
 }
