+++ src/main/resources/liquibase/harness/diff/changelogs/mssql2022_to_mssql2017.xml
@@ -134,26 +134,26 @@
     <!-- PACKAGE -->
     <changeSet id="13" author="as">
         <pro:createPackage encoding="UTF-8"
-                           packageName="Test package">CREATE FUNCTION test_function()
+                           packageName="Test package">CREATE FUNCTION test_function1()
             RETURNS VARCHAR(20)
             BEGIN
             RETURN 'Hello';
             END</pro:createPackage>
         <rollback>
-            <pro:dropFunction functionName="test_function"/>
+            <pro:dropFunction functionName="test_function1"/>
         </rollback>
     </changeSet>
 
     <!-- PACKAGE BODY -->
     <changeSet id="14" author="as">
         <pro:createPackageBody encoding="UTF-8"
-                               packageBodyName="Test package">CREATE FUNCTION test_function()
+                               packageBodyName="Test package">CREATE FUNCTION test_function2()
             RETURNS VARCHAR(20)
             BEGIN
             RETURN 'Hello';
             END</pro:createPackageBody>
         <rollback>
-            <pro:dropFunction functionName="test_function"/>
+            <pro:dropFunction functionName="test_function2"/>
         </rollback>
     </changeSet>
