+++ src/main/resources/liquibase/harness/diff/diffDatabases.yml
@@ -34,12 +34,13 @@ references:
     referenceDatabaseName: mssql
     referenceDatabaseVersion: 2022
 
-  - targetDatabaseName: mysql
-    targetDatabaseVersion: 5.7
-    referenceDatabaseName: mysql
-    referenceDatabaseVersion: 8.4
-
+#MySQL should be always latest
   - targetDatabaseName: mysql
     targetDatabaseVersion: 5.6
     referenceDatabaseName: mysql
     referenceDatabaseVersion: 8
+
+  - targetDatabaseName: mysql
+    targetDatabaseVersion: 5.7
+    referenceDatabaseName: mysql
+    referenceDatabaseVersion: 8.4
