+++ src/test/groovy/liquibase/harness/DiffChangelogTest.groovy
@@ -1,0 +1,6 @@
+package liquibase.harness
+
+import liquibase.harness.diff.DiffChangelogTests;
+
+class DiffChangelogTest extends DiffChangelogTests {
+}
